pyproject.toml
[tool.poetry]
name = "loyalty"
version = "0.1.0"
description = ""
authors = ["cmrd-a <cmrd.a@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.51.1"
grpcio-tools = "^1.51.1"
alembic = "1.9.0"
pydantic = "^1.10.2"
grpcio-reflection = "^1.51.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0b4", allow-prereleases = true}

[tool.black]
line-length = 120
extend_exclude = ".*pb2.*.py"

[tool.ruff]
line-length = 120
extend-exclude = ["*pb2*.py", "*.pyi"]
ignore = ["E999"]


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
ruff = "^0.0.187"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
